Tested Rest Api for Category

+++ MusicStore/src/main/java/com/musicstore/controller/api/CategoryRestController.java
@@ -1,6 +1,7 @@
 package com.musicstore.controller.api;
 
 import java.util.List;
+import java.util.Map;
 import java.util.ArrayList; 
 import java.util.Optional;
 
@@ -14,11 +15,22 @@ import org.springframework.web.server.ResponseStatusException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
+import com.musicstore.model.CartBean;
 import com.musicstore.model.CategoryBean;
+import com.musicstore.model.WebUserBean;
+import com.musicstore.service.DbAdminService;
 import com.musicstore.service.DbCategoryService;
+import com.musicstore.service.DbWebUserService;
+import com.musicstore.utility.Utility;
 
 @RestController
 public class CategoryRestController {
+
+	@Autowired
+	private DbAdminService adminService; 
+	
+	@Autowired
+	private DbWebUserService webuserService;
 	
 	@Autowired
 	private DbCategoryService categoryService; 
@@ -26,12 +38,17 @@ public class CategoryRestController {
 	public CategoryRestController() {}
 	
 	@RequestMapping("/musicstore/api/category")
-	public Iterable<CategoryBean> getAll(){
+	public Iterable<CategoryBean> getAll(@RequestBody WebUserBean b){
+		if(!adminService.isAdmin(b))
+		{
+			throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "request by not an admin");
+		}
 		return categoryService.getAll();
 	}
 	
 	@RequestMapping("/musicstore/api/category/{id}")
 	public CategoryBean getById(@PathVariable int id){
+		/*EVERYONE SHOULD BE ABLE TO GET INFOS ABOUT SPECIFIC CATEGORY*/
 		Optional<CategoryBean> category = categoryService.getById(id);
 		if(category.isEmpty()){
 			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "item not found");
@@ -40,14 +57,23 @@ public class CategoryRestController {
 	}
 
 	@RequestMapping(value  ="/musicstore/api/category", method = RequestMethod.POST)
-	public CategoryBean create(@RequestBody CategoryBean p) {
-		return categoryService.create(p);
+	public CategoryBean create(@RequestBody Map<String, Map<String,String>> map) {
+		CategoryBean cb = Utility.categoryDeMap(map.get("topost"));
+		WebUserBean b = Utility.webuserDeMap(map.get("authorized"));
+		if(!adminService.isAdmin(b)){
+			throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "request by not an admin");
+		}
+		return categoryService.create(cb);
 	}
 	
 	@RequestMapping(value  ="/musicstore/api/category/{id}", method = RequestMethod.PUT)
-	public CategoryBean update(@PathVariable int id, @RequestBody CategoryBean p) {
-		
-		Optional<CategoryBean> updatedCategory= categoryService.update(id, p);
+	public CategoryBean update(@PathVariable int id,@RequestBody Map<String, Map<String,String>> map) {
+		CategoryBean cb = Utility.categoryDeMap(map.get("toput"));
+		WebUserBean b = Utility.webuserDeMap(map.get("authorized"));
+		if(!adminService.isAdmin(b)){
+			throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "request by not an admin");
+		}
+		Optional<CategoryBean> updatedCategory= categoryService.update(id, cb);
 		if (updatedCategory.isEmpty())
 		{
 			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "item not found");
@@ -56,11 +82,14 @@ public class CategoryRestController {
 	}
 	
 	@RequestMapping(value  ="/musicstore/api/category/{id}", method = RequestMethod.DELETE)
-	public void delete(@PathVariable int id) {
+	public void delete(@PathVariable int id, @RequestBody WebUserBean b) {
+		if(!adminService.isAdmin(b)) {
+			throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "request by not an admin");
+		}
 		Boolean isDeleted = categoryService.delete(id);
 		if (isDeleted==false)
 		{
-			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "item not found");
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Category not found");
 		}
 	}
